# backend/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/public/rooms/{slug}": {
      "get": {
        "operationId": "CompatController_roomDetail",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Compat"
        ]
      }
    },
...
    "/api/partner/rooms/{id}": {
      "get": {
        "operationId": "CompatController_partnerRoom",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Compat"
        ]
      },
      "patch": {
        "operationId": "CompatController_updatePartnerRoom",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Compat"
        ]
      },
      "delete": {
        "operationId": "CompatController_deletePartnerRoom",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Compat"
        ]
      }
    },
    "/api/partner/rooms/{id}/request-update": {
      "patch": {
        "operationId": "CompatController_requestUpdateRoom",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Compat"
        ]
      }
    },
    "/api/partner/rooms/{id}/request-delete": {
      "delete": {
        "operationId": "CompatController_requestDeleteRoom",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Compat"
        ]
      }
    },
    "/api/partner/rooms/{id}/request-restore": {
      "post": {
        "operationId": "CompatController_requestRestoreRoom",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Compat"
        ]
      }
    },
...
    "/api/admin/admins/{id}": {
      "patch": {
        "operationId": "CompatController_updateAdmin",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Compat"
        ]
      },
      "delete": {
        "operationId": "CompatController_deleteAdmin",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Compat"
        ]
      }
    },
    "/api/admin/partners": {
      "get": {
        "operationId": "CompatController_partners",
        "parameters": [
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Compat"
        ]
      }
    },
    "/api/admin/partners/{id}": {
      "patch": {
        "operationId": "CompatController_updatePartner",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Compat"
        ]
      },
      "delete": {
        "operationId": "CompatController_deletePartner",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Compat"
        ]
      }
    },
    "/api/admin/partners/{id}/approve": {
      "post": {
        "operationId": "CompatController_approvePartner",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Compat"
        ]
      }
    },
    "/api/admin/partners/{id}/reject": {
      "post": {
        "operationId": "CompatController_rejectPartner",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Compat"
        ]
      }
    },
    "/api/admin/partners/{id}/rooms": {
      "get": {
        "operationId": "CompatController_partnerRoomsForAdmin",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Compat"
        ]
      }
    },
...
    "/api/admin/rooms/{id}/approve": {
      "post": {
        "operationId": "CompatController_approveRoom",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Compat"
        ]
      }
    },
    "/api/admin/rooms/{id}/reject": {
      "post": {
        "operationId": "CompatController_rejectRoom",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Compat"
        ]
      }
    },
    "/api/admin/rooms/{id}": {
      "patch": {
        "operationId": "CompatController_updateRoom",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Compat"
        ]
      },
      "delete": {
        "operationId": "CompatController_deleteRoom",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Compat"
        ]
      }
    },
...
    "/api/admin/bookings/{id}/cancel": {
      "post": {
        "operationId": "CompatController_adminCancelBooking",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Compat"
        ]
      }
    },
    "/api/admin/bookings/{id}/mark-paid": {
      "post": {
        "operationId": "CompatController_adminMarkBookingPaid",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Compat"
        ]
      }
    },
    "/api/admin/bookings/{id}/reject-cancel": {
      "post": {
        "operationId": "CompatController_adminRejectBookingCancel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Compat"
        ]
      }
    },
...
    "/api/partner/bookings/{id}/{action}": {
      "post": {
        "operationId": "CompatController_partnerBookingAction",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "action",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Compat"
        ]
      }
    },
...
    "/api/bookings/{id}/status": {
      "get": {
        "operationId": "CompatController_bookingStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Compat"
        ]
      }
    },
...
  },
  "info": {
    "title": "NoWayHome API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}